refactor: log-manager-files
