+++ .changeset/brown-moons-poke.md
@@ -1,0 +1,5 @@
+---
+'@koopjs/winnow': patch
+---
+
+- rename log-manager files
