+++ .changeset/modern-seals-prove.md
@@ -1,0 +1,5 @@
+---
+'@koopjs/featureserver': patch
+---
+
+- rename log-manager files
