+++ packages/featureserver/src/helpers/feature-layer-metadata.js
@@ -6,7 +6,7 @@ const {
   PolygonRenderer
 } = require('./renderers');
 const { calculateBounds } = require('@terraformer/spatial');
-const logManager = require('../logger');
+const logManager = require('../log-manager');
 const getSpatialReference = require('./get-spatial-reference');
 const getGeometryTypeFromGeojson = require('./get-geometry-type-from-geojson');
 const normalizeExtent = require('./normalize-extent');
