+++ packages/featureserver/src/helpers/get-geometry-type-from-geojson.js
@@ -1,5 +1,5 @@
 const _ = require('lodash');
-const logManager = require('../logger');
+const logManager = require('../log-manager');
 
 const esriLookup = {
   Point: 'esriGeometryPoint',
