+++ packages/featureserver/src/route.js
@@ -1,7 +1,7 @@
 const _ = require('lodash');
 const layerInfo = require('./layer-metadata');
 const query = require('./query');
-const logManager = require('./logger');
+const logManager = require('./log-manager');
 const queryRelatedRecords = require('./queryRelatedRecords.js');
 const generateRenderer = require('./generate-renderer');
 const restInfo = require('./rest-info-route-handler');
