+++ packages/winnow/src/normalize-query-options/source-data-spatial-reference.js
@@ -1,6 +1,6 @@
 const normalizeSpatialReference = require('./spatial-reference');
 const { getCollectionCrs } = require('./helpers');
-const logManager = require('../logger');
+const logManager = require('../log-manager');
 
 function normalizeSourceDataSpatialReference ({ inputCrs, sourceSR, collection } = {}) {
   const sourceDataSpatialReference = inputCrs || sourceSR || getCollectionCrs(collection);
