+++ packages/winnow/src/normalize-query-options/spatial-reference.js
@@ -1,7 +1,7 @@
 const esriProjCodes = require('@esri/proj-codes');
 const Joi = require('joi');
 const wktParser = require('wkt-parser');
-const logManager = require('../logger');
+const logManager = require('../log-manager');
 const PROJ4_WKIDS = [4326, 4269, 3857, 3785, 900913, 102113];
 const wktLookup = new Map();
 const schema = Joi.alternatives(
